Fix FAST lint warnings (#21)
# Pull Request

## 📖 Description

When you create a PR in this repo, GitHub displays a bunch of lint
warnings for files that aren't part of the PR. I would like to get rid
of this noise. For example: https://github.com/ni/fast/pull/19/files

### 🎫 Issues

None

## 👩‍💻 Reviewer Notes

I manually fixed several warnings. Most of them were unused variables or
conditional statements with side effects. There was only one functional
change, which I'll call out in a comment.

## 📑 Test Plan

Relying on pipeline

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [x] I have included a change request file using `$ npm run change`
- [ ] I have added tests for my changes.
- [ ] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.
- [ ] I have read the
[CONTRIBUTING](https://github.com/microsoft/fast/blob/main/CONTRIBUTING.md)
documentation and followed the
[standards](https://github.com/microsoft/fast/blob/main/CODE_OF_CONDUCT.md#our-standards)
for this project.

## ⏭ Next Steps

<!---
If there is relevant follow-up work to this PR, please list any existing
issues or provide brief descriptions of what you would like to do next.
-->