Update utilities with path resolving logic that removes ../ (#7163)
# Pull Request

## 📖 Description

This change assumes non-context containing paths refer to the root component, it also removes partial and apply as part of the templating syntax.

## 👩‍💻 Reviewer Notes

This change is going into a feature branch, the next check in should include removal of path caching in Observer Map, and instead rely on the schema generated for context.

## ✅ Checklist

### General

- [x] I have included a change request file using `$ npm run change`
- [x] I have added tests for my changes.
- [x] I have tested my changes.
- [x] I have updated the project documentation to reflect my changes.
- [x] I have read the [CONTRIBUTING](https://github.com/microsoft/fast/blob/main/CONTRIBUTING.md) documentation and followed the [standards](https://github.com/microsoft/fast/blob/main/CODE_OF_CONDUCT.md#our-standards) for this project.