remove shadowOptions from TemplateElement.options() (#7160)
<!---
Thanks for filing a pull request 😄 ! Before you submit, please read the following:

Search open/closed issues before submitting. Someone may have pushed the same thing before!

Provide a summary of your changes in the title field above.
-->

# Pull Request

## 📖 Description

Remove the option of adding `shadowOptions` in `TemplateElement.options()`. For defining `shadowOptions`, do so in `defineAsync()` instead, similar to how `define()` or `compose()` works.

### 🎫 Issues

Requiring defining `shadowOptions` on `TemplateElement.options()` means the component users have to be aware of the component's definitions, which it's possible, it doesn't feel nature, and makes it hard for component author to control the `defineAsync()` call.

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [x] I have included a change request file using `$ npm run change`
- [x] I have added tests for my changes.
- [x] I have tested my changes.
- [x] I have updated the project documentation to reflect my changes.
- [x] I have read the [CONTRIBUTING](https://github.com/microsoft/fast/blob/main/CONTRIBUTING.md) documentation and followed the [standards](https://github.com/microsoft/fast/blob/main/CODE_OF_CONDUCT.md#our-standards) for this project.