Create a more elegant solution when setting objects to avoid Object.assign (#7190)
# Pull Request

## 📖 Description

Currently Object.assign must be used when setting objects which have already been defined, this is un-ergonomic as it presupposes knowledge from a developer that a proxy is being used on this object. To circumvent that we now check in the _Changed function whether a proxy is assigned to the object, if not we can assume the user is setting the object and re-proxy the object.

## ✅ Checklist

### General

- [x] I have included a change request file using `$ npm run change`
- [x] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.
- [x] I have read the [CONTRIBUTING](https://github.com/microsoft/fast/blob/main/CONTRIBUTING.md) documentation and followed the [standards](https://github.com/microsoft/fast/blob/main/CODE_OF_CONDUCT.md#our-standards) for this project.